# swagger/admins.yml
openapi: 3.0.0    
paths:
  /admins/dashboard:
    get:
      tags:
        - Admins
      summary: Zusammenfassung aller User, Events, Violations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/admins:
    get:
      tags:
        - Admins
      summary: Alle Admins abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/users:
    get:
      tags:
        - Admins
      summary: Alle Benutzer abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/events:
    get:
      tags:
        - Admins
      summary: Alle Events abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/violations:
    get:
      tags:
        - Admins
      summary: Alle Violations abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/comments:
    get:
      tags:
        - Admins
      summary: Alle Comments abrufen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/users/role:
    put:
      tags:
        - Admins
      summary: Benutzerrolle ändern zu "user" oder "admin"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 4
                role: admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/violations/status:
    put:
      tags:
        - Admins
      summary: Violations status ändern "resolved" oder "pending"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id: 1
                status: resolved
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/users/delete:
    delete:
      tags:
        - Admins
      summary: Löschen eines Users anhand der UserId
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id: 5
      responses:
        '200':
          description: Erfolgreiches Löschen des Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted
        '404':
          description: User nicht gefunden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Serverfehler
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /admins/events/delete:
    delete:
      tags:
        - Admins
      summary: Löschen eines Events anhand der EventId
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id: 3
      responses:
        '200':
          description: Erfolgreiches Löschen des Events
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event deleted
        '404':
          description: Event nicht gefunden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event not found
        '500':
          description: Serverfehler
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /admins/comments/delete:
    delete:
      tags:
        - Admins
      summary: Löschen eines Kommentars anhand der Kommentar-ID
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id: 7
      responses:
        '200':
          description: Erfolgreiches Löschen des Kommentars
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Comment deleted
        '404':
          description: Kommentar nicht gefunden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Comment not found
        '500':
          description: Serverfehler
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error